% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_debug}
\alias{fes_debug}
\title{Enables debug print statements showing the process flow of Rfes2014}
\usage{
fes_debug(enable)
}
\arguments{
\item{enable}{logical; TRUE to enable, FALSE to disable}
}
\value{
None
}
\description{
Enables debug print statements showing the process flow of Rfes2014
}
\examples{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
# Created FES handle: 9c5b5020
fes_calculate(hdl, 50.10, -2.40, 1570536000)
# Calculating for: 50.100000 -2.400000 25482.500000
# Result: -24.965188 0.210749 4
fes_delete(hdl)
# Destroying FES handle: 9c5b5020

\dontrun{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
# Created FES handle: 8ca5f830
q()
# Finalizing handle: 8ca5f830
# Destroying FES handle: 8ca5f830
}

}
