% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_debugCalculate}
\alias{fes_debugCalculate}
\title{Enables debug print statements showing the calculation flow of Rfes2014}
\usage{
fes_debugCalculate(enable)
}
\arguments{
\item{enable}{logical; TRUE to enable, FALSE to disable}
}
\value{
None
}
\description{
Enables debug print statements showing the calculation flow of Rfes2014
}
\examples{
\dontrun{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
fes_calculate(hdl, 50.10, -2.40, 1570536000)
# UseMethod Type: double
# Using generic: fes_calculate.double
# Calculating for: 50.100000 -2.400000 25482.500000
# Result: -24.965188 0.210749 4
# Calculating for: 50.100000 -2.400000 25482.500000
# Result: -24.965188 0.210749 4
fes_delete(hdl)
}

\dontrun{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
fes_calculate(hdl, c(50.10, 50.10), c(-2.50, -2.40), c(1570536000, 1570536000))
# UseMethod Type: double; Class: numeric; Length: 3
# Using generic: fes_calculate.list
# Processing 2 rows
#   [0] Calculating for: 50.100000 -2.500000 25482.500000
#   [0] Result: -22.001810 0.227711 4
#   [1] Calculating for: 50.100000 -2.400000 25482.500000
#   [1] Result: -24.965188 0.210749 4
fes_delete(hdl)
}

}
