% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_debugHandles}
\alias{fes_debugHandles}
\title{Enables FES2014 handle debugging}
\usage{
fes_debugHandles(enable)
}
\arguments{
\item{enable}{logical; TRUE to enable, FALSE to disable}
}
\value{
None
}
\description{
Enables FES2014 handle debugging
}
\examples{
\dontrun{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
# Created FES handle: 9c5b5020
fes_calculate(hdl, 50.10, -2.40, 1570536000)
fes_delete(hdl)
# Destroying FES handle: 9c5b5020
}

\dontrun{
fes_debug(TRUE)
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
# Created FES handle: 8ca5f830
q()
# Finalizing handle: 8ca5f830
# Destroying FES handle: 8ca5f830
}

}
