% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_delete}
\alias{fes_delete}
\title{Cleans up a FES2014 instance}
\usage{
fes_delete(hdl)
}
\arguments{
\item{hdl}{<extptr>; a FES2014 handle}
}
\value{
None
}
\description{
Cleans up a FES2014 instance
}
\examples{
\dontrun{
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
fes_delete(hdl)
}

}
\seealso{
\code{\link{fes_new}}
}
