% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_isDebugCalculateEnabled}
\alias{fes_isDebugCalculateEnabled}
\title{Checks if calculation debug output is enabled}
\usage{
fes_isDebugCalculateEnabled()
}
\value{
logical; TRUE if debug is enabled, FALSE if debug is disabled
}
\description{
Checks if calculation debug output is enabled
}
\examples{
fes_isDebugCalculateEnabled()
fes_debug(TRUE)
fes_isDebugCalculateEnabled()

}
