% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_isDebugHandlesEnabled}
\alias{fes_isDebugHandlesEnabled}
\title{Checks if debugging handles is enabled}
\usage{
fes_isDebugHandlesEnabled()
}
\value{
logical; TRUE if debug is enabled, FALSE if debug is disabled
}
\description{
Checks if debugging handles is enabled
}
\examples{
fes_isDebugHandlesEnabled()
fes_debugHandles(TRUE)
fes_isDebugHandlesEnabled()

}
