% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014.R
\name{fes_new}
\alias{fes_new}
\title{Create a handle to a new FES2014 instance}
\usage{
fes_new(tide_type, access_mode, ini_file)
}
\arguments{
\item{tide_type}{string; one of "ocean", "tide" or "radial"}

\item{access_mode}{string; one of "io" or "memory"}

\item{ini_file}{string; path to a fes2014 configuration file}
}
\value{
<extpointer> A handle to a FES2014 instance
}
\description{
FES2014 handles are used in \code{\link{fes_calculate}} and can either be
cleaned up manually with \code{\link{fes_delete}} or are finalized
automatically when the R session is destroyed.
}
\details{
DO NOT SAVE THE HANDLE IN RDATA FILES
}
\examples{
\dontrun{
hdl <- fes_new("ocean", "memory", "data/fes2014/eastward_velocity.ini")
}

}
\seealso{
\code{\link{fes_delete}}
}
