% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfes2014-package.R
\docType{package}
\name{Rfes2014-package}
\alias{Rfes2014}
\alias{Rfes2014-package}
\title{Rfes2014: R wrapper for the FES2014 library provided by Aviso}
\description{
R wrapper for the FES2014 c library provided by Aviso, see https://bitbucket.org/cnes_aviso/fes/src
}
\details{
The Rfes2014 package provides an R wrapper around the FES2014 library made
available by Aviso.

There are 3 main functions to Rfes2014:
\itemize{
  \item \code{\link{fes_new}}: Create a new FES2014 instance
  \item \code{\link{fes_calculate}}: Request FES2014 to do a calculation
  \item \code{\link{fes_delete}}: Clean up the FES2014 instance
}
}
\references{
AVSDev: \url{https://avsdev.uk/R/Rfes}

FES2014: \url{https://bitbucket.org/cnes_aviso/fes}

netCDF: \url{https://github.com/Unidata/netcdf-c}
}
\author{
Craig Williams, AVS Developments Ltd \email{craig@avsdev.uk}
}
